requirements.txt
attrs==24.2.0
autopep8==2.3.1
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
packaging==24.1
pycodestyle==2.12.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
termcolor==2.4.0
